/Akshay

+++ src/main/java/com/ats/tril/model/doc/POReportDetail.java
@@ -24,15 +24,15 @@ public class POReportDetail {
 
 	private String itemUom;
 
-	private int itemQty;
+	private float itemQty;
 
 	private float itemRate;
 
-	private int mrnQty;
+	private float mrnQty;
 
-	private int pendingQty;
+	private float pendingQty;
 
-	private int indedQty;
+	private float indedQty;
 
 	private float discPer;
 
@@ -141,11 +141,11 @@ public class POReportDetail {
 		this.itemUom = itemUom;
 	}
 
-	public int getItemQty() {
+	public float getItemQty() {
 		return itemQty;
 	}
 
-	public void setItemQty(int itemQty) {
+	public void setItemQty(float itemQty) {
 		this.itemQty = itemQty;
 	}
 
@@ -157,27 +157,27 @@ public class POReportDetail {
 		this.itemRate = itemRate;
 	}
 
-	public int getMrnQty() {
+	public float getMrnQty() {
 		return mrnQty;
 	}
 
-	public void setMrnQty(int mrnQty) {
+	public void setMrnQty(float mrnQty) {
 		this.mrnQty = mrnQty;
 	}
 
-	public int getPendingQty() {
+	public float getPendingQty() {
 		return pendingQty;
 	}
 
-	public void setPendingQty(int pendingQty) {
+	public void setPendingQty(float pendingQty) {
 		this.pendingQty = pendingQty;
 	}
 
-	public int getIndedQty() {
+	public float getIndedQty() {
 		return indedQty;
 	}
 
-	public void setIndedQty(int indedQty) {
+	public void setIndedQty(float indedQty) {
 		this.indedQty = indedQty;
 	}
